OSGI-204 OSGI-215 OSGI-219 OSGI-216 OSGI-200 OSGI-257

+++ core/src/main/java/org/springframework/osgi/internal/service/collection/OsgiServiceCollection.java
@@ -191,7 +191,7 @@ public class OsgiServiceCollection implements Collection, InitializingBean, Impo
 
 	private final BundleContext context;
 
-	private int contextClassLoader = ReferenceClassLoadingOptions.CLIENT;
+	private int contextClassLoader = ReferenceClassLoadingOptions.CLIENT.shortValue();
 
 	protected final ClassLoader classLoader;
